+++ tests/test_version.py
@@ -6,9 +6,7 @@ import wordseg
 
 def test_version_number_match_with_changelog():
     """__version__ and CHANGELOG.md match for the latest version number."""
-    repo_dir = os.path.dirname(
-        os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-    )
+    repo_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
     changelog = open(os.path.join(repo_dir, "CHANGELOG.md")).read()
     # latest version number in changelog = the 1st occurrence of '[x.y.z]'
     version_in_changelog = (